Dev

See merge request main-dev-group/league-core!318

+++ src/main/java/com/freetonleague/core/domain/dto/RouletteStatsDto.java
@@ -13,7 +13,7 @@ import java.util.List;
 @AllArgsConstructor
 public class RouletteStatsDto {
 
-    private Integer gamesPlayedToday;
+    private Long gamesPlayedToday;
     private Long tonAmountForToday;
     private Long tonAmountForAllTime;
